refactor(terminal-themes): share the resolver and the mobile projection
The terminal-theme resolver and the mobile wire projection lived in the
renderer, so the main process had no way to answer "what palette is this host
configured for" — which is the whole reason a phone paired to `orca serve` or an
SSH host renders a hardcoded fallback. Extracting rather than reimplementing is
what makes the follow-up stamp byte-for-byte identical to the renderer's output
by construction, instead of by inspection.

resolveEffectiveTerminalAppearance loses its `systemPrefersDark = getSystemPrefersDark()`
default argument: src/shared must stay DOM-free (the suite runs under
environment: 'node', and Metro consumes the tree too). All 17 call sites already
passed the value explicitly, so no caller changes.

EffectiveTerminalAppearance.theme is typed TerminalColorOverrides instead of
xterm's ITheme. Same 22 keys plus `bold`, already the mobile wire shape, and it
keeps @xterm/xterm out of src/shared entirely.

terminal-theme.ts keeps a re-export shim rather than rewriting 22 import sites
and the 4 vi.mock factories that key on its module specifier: 0 changed
importers, 0 changed mocks.

resolveRuntimeMobileTerminalTheme takes a Partial<Pick<GlobalSettings, ...>>
instead of AppState so main can pass its narrowed store settings. The three
normalizations are all no-ops on a full GlobalSettings: `?? ''` on the two theme
names and the two divider colors hits fallbacks the resolver already applies,
and `terminalUseSeparateLightTheme !== false` preserves the shipped `true`
default against a store type that makes the key optional.

sync-runtime-graph.test.ts is unedited and still pins the full projection —
including the rgba(248, 248, 248, 0.8) / rgba(32, 32, 32, 0.5) opacity literals
and the 'system' -> mode cases. Its passing unchanged is the proof no value
moved.
