+++ src/renderer/src/lib/terminal-theme.ts
@@ -7,26 +7,24 @@ import type { GlobalSettings } from '../../../shared/types'
 import {
   makeCustomTerminalThemeSelection,
   normalizeTerminalCustomThemes,
-  parseCustomTerminalThemeSelection,
   terminalCustomThemeToXtermTheme,
   type TerminalCustomTheme
 } from '../../../shared/terminal-custom-themes'
 
 export { BUILTIN_TERMINAL_THEME_NAMES }
 
-export const DEFAULT_TERMINAL_THEME_DARK = 'Ghostty Default Style Dark'
-export const DEFAULT_TERMINAL_THEME_LIGHT = 'Builtin Tango Light'
-export const DEFAULT_TERMINAL_DIVIDER_DARK = '#3f3f46'
-const DEFAULT_TERMINAL_DIVIDER_LIGHT = '#d4d4d8'
-
-export type EffectiveTerminalAppearance = {
-  mode: 'dark' | 'light'
-  sourceTheme: 'system' | 'dark' | 'light'
-  themeName: string
-  dividerColor: string
-  theme: ITheme | null
-  systemPrefersDark: boolean
-}
+// Why: the resolver is shared with the main process (headless mobile tabs); this
+// module stays the renderer's single import surface.
+export {
+  DEFAULT_TERMINAL_DIVIDER_DARK,
+  DEFAULT_TERMINAL_THEME_DARK,
+  DEFAULT_TERMINAL_THEME_LIGHT,
+  getTerminalTheme,
+  getTerminalThemePreview,
+  normalizeColor,
+  resolveEffectiveTerminalAppearance,
+  type EffectiveTerminalAppearance
+} from '../../../shared/terminal-theme-resolution'
 
 export type TerminalThemeOption = {
   value: string
@@ -46,46 +44,6 @@ export function getSystemPrefersDark(): boolean {
 
 export function getBuiltinTheme(name: string): ITheme | null {
   return getBuiltinTerminalThemePalette(name)
-}
-
-function findCustomTheme(
-  settings: Pick<GlobalSettings, 'terminalCustomThemes'> | undefined,
-  selection: string
-): TerminalCustomTheme | null {
-  const customId = parseCustomTerminalThemeSelection(selection)
-  if (!customId || !settings) {
-    return null
-  }
-  return (
-    normalizeTerminalCustomThemes(settings.terminalCustomThemes).find(
-      (theme) => theme.id === customId
-    ) ?? null
-  )
-}
-
-export function getTerminalTheme(
-  settings: Pick<GlobalSettings, 'terminalCustomThemes'> | undefined,
-  selection: string
-): ITheme | null {
-  const customTheme = findCustomTheme(settings, selection)
-  if (customTheme) {
-    return terminalCustomThemeToXtermTheme(customTheme)
-  }
-  return getBuiltinTerminalThemePalette(selection)
-}
-
-export function getTerminalThemePreview(
-  name: string,
-  settings?: Pick<GlobalSettings, 'terminalCustomThemes'>,
-  fallbackMode: 'dark' | 'light' = 'dark'
-): ITheme | null {
-  const theme = getTerminalTheme(settings, name)
-  if (theme) {
-    return theme
-  }
-  return getBuiltinTerminalThemePalette(
-    fallbackMode === 'light' ? DEFAULT_TERMINAL_THEME_LIGHT : DEFAULT_TERMINAL_THEME_DARK
-  )
 }
 
 export function getAvailableTerminalThemeOptions(
@@ -109,47 +67,6 @@ export function getAvailableTerminalThemeOptions(
     })
   )
   return [...builtinOptions, ...customOptions]
-}
-
-export function resolveEffectiveTerminalAppearance(
-  settings: Pick<
-    GlobalSettings,
-    | 'theme'
-    | 'terminalThemeDark'
-    | 'terminalDividerColorDark'
-    | 'terminalUseSeparateLightTheme'
-    | 'terminalThemeLight'
-    | 'terminalCustomThemes'
-    | 'terminalDividerColorLight'
-  >,
-  systemPrefersDark = getSystemPrefersDark()
-): EffectiveTerminalAppearance {
-  const sourceTheme =
-    settings.theme === 'system' ? (systemPrefersDark ? 'dark' : 'light') : settings.theme
-  const useLightVariant = sourceTheme === 'light' && settings.terminalUseSeparateLightTheme
-  const themeName = useLightVariant
-    ? settings.terminalThemeLight || DEFAULT_TERMINAL_THEME_LIGHT
-    : settings.terminalThemeDark || DEFAULT_TERMINAL_THEME_DARK
-  const dividerColor = useLightVariant
-    ? normalizeColor(settings.terminalDividerColorLight, DEFAULT_TERMINAL_DIVIDER_LIGHT)
-    : normalizeColor(settings.terminalDividerColorDark, DEFAULT_TERMINAL_DIVIDER_DARK)
-
-  return {
-    mode: sourceTheme,
-    sourceTheme: settings.theme,
-    themeName,
-    dividerColor,
-    theme: getTerminalThemePreview(themeName, settings, useLightVariant ? 'light' : 'dark'),
-    systemPrefersDark
-  }
-}
-
-export function normalizeColor(value: string | undefined, fallback: string): string {
-  const trimmed = value?.trim()
-  if (!trimmed) {
-    return fallback
-  }
-  return trimmed
 }
 
 export function clampNumber(value: number, min: number, max: number): number {
