+++ src/renderer/src/runtime/sync-runtime-graph.ts
@@ -9,8 +9,9 @@ import { getEagerPtyBufferHandle } from '@/components/terminal-pane/pty-dispatch
 import { createBrowserUuid } from '@/lib/browser-uuid'
 import type { PaneManager } from '@/lib/pane-manager/pane-manager'
 import { resolveLeafIdForManager } from '@/lib/pane-manager/pane-key-resolution'
-import { getSystemPrefersDark, resolveEffectiveTerminalAppearance } from '@/lib/terminal-theme'
+import { getSystemPrefersDark } from '@/lib/terminal-theme'
 import { sanitizeTerminalLayoutPaneTitles } from '@/lib/terminal-pane-title-sanitization'
+import { resolveRuntimeMobileTerminalTheme } from '../../../shared/runtime-mobile-terminal-theme'
 import type { AppState } from '@/store/types'
 import type {
   RuntimeMobileSessionBrowserTab,
@@ -18,8 +19,8 @@ import type {
   RuntimeMobileSessionMarkdownTab,
   RuntimeMobileSessionTabGroup,
   RuntimeMobileSessionSnapshotTab,
-  RuntimeMobileTerminalTheme,
   RuntimeMobileSessionTabsSnapshot,
+  RuntimeMobileTerminalTheme,
   RuntimeSyncWindowGraph
 } from '../../../shared/runtime-types'
 import { isTerminalLeafId, makePaneKey, parsePaneKey } from '../../../shared/stable-pane-id'
@@ -459,7 +460,9 @@ export function getRuntimeMobileSessionSyncKey(
       state.settings === previousState.settings &&
       previousKey.systemPrefersDark === terminalThemeSystemPrefersDark
         ? previousKey.terminalThemeProjection
-        : JSON.stringify(resolveMobileTerminalTheme(state, systemPrefersDark) ?? null),
+        : JSON.stringify(
+            resolveRuntimeMobileTerminalTheme(state.settings, systemPrefersDark) ?? null
+          ),
     // Why: background title ticks churn many times/sec; reuse unchanged projections so they don't rescan all tabs, files, and drafts.
     tabsProjection:
       canReusePrevious && state.tabsByWorktree === previousState.tabsByWorktree
@@ -942,7 +945,7 @@ function getMobileTerminalTheme(
   ) {
     return cachedMobileTerminalTheme
   }
-  cachedMobileTerminalTheme = resolveMobileTerminalTheme(state, systemPrefersDark)
+  cachedMobileTerminalTheme = resolveRuntimeMobileTerminalTheme(state.settings, systemPrefersDark)
   cachedMobileTerminalThemeSettings = state.settings
   cachedMobileTerminalThemeSystemPrefersDark = systemPrefersDark
   hasCachedMobileTerminalTheme = true
@@ -1735,58 +1738,6 @@ function getEditorDraftVersionByFileId(
 
 function mobileTerminalSurfaceId(parentTabId: string, leafId: string): string {
   return `${parentTabId}::${leafId}`
-}
-
-function hexToRgba(hex: string, alpha: number): string {
-  let clean = hex.replace('#', '')
-  if (clean.length === 3) {
-    clean = clean
-      .split('')
-      .map((c) => c + c)
-      .join('')
-  }
-  const r = Number.parseInt(clean.slice(0, 2), 16)
-  const g = Number.parseInt(clean.slice(2, 4), 16)
-  const b = Number.parseInt(clean.slice(4, 6), 16)
-  return `rgba(${r}, ${g}, ${b}, ${alpha})`
-}
-
-function isHexColor(value: string | undefined): value is string {
-  return typeof value === 'string' && /^#(?:[0-9a-fA-F]{3}|[0-9a-fA-F]{6})$/.test(value)
-}
-
-function resolveMobileTerminalTheme(
-  state: AppState,
-  systemPrefersDark: boolean
-): RuntimeMobileTerminalTheme | undefined {
-  const settings = state.settings
-  if (!settings) {
-    return undefined
-  }
-  const appearance = resolveEffectiveTerminalAppearance(settings, systemPrefersDark)
-  const resolvedTheme = appearance.theme
-    ? { ...appearance.theme, ...settings.terminalColorOverrides }
-    : undefined
-  if (!resolvedTheme) {
-    return undefined
-  }
-  if (settings.terminalBackgroundOpacity !== undefined && isHexColor(resolvedTheme.background)) {
-    resolvedTheme.background = hexToRgba(
-      resolvedTheme.background,
-      settings.terminalBackgroundOpacity
-    )
-  }
-  if (settings.terminalCursorOpacity !== undefined && isHexColor(resolvedTheme.cursor)) {
-    resolvedTheme.cursor = hexToRgba(resolvedTheme.cursor, settings.terminalCursorOpacity)
-  }
-
-  const theme: Record<string, string> = {}
-  for (const [key, value] of Object.entries(resolvedTheme)) {
-    if (typeof value === 'string') {
-      theme[key] = value
-    }
-  }
-  return { mode: appearance.mode, theme: theme as RuntimeMobileTerminalTheme['theme'] }
 }
 
 function getRuntimeLeafIdsForTerminal(
